JetsonNano_AI/requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
pyserial>=3.5
ultralytics>=8.0.0
scipy>=1.7.0
pyrtlsdr>=0.2.93
